Separate BookUpdateService to BookHistoryUpdateService and HistoryService to many Services
